requirements.txt
python-decouple==3.8
requests==2.32.5
terminaltables==3.1.10